list_shows y list_episodes

+++ initdb.sql
@@ -2,8 +2,8 @@ CREATE TABLE shows (
     id INTEGER PRIMARY KEY,
     name VARCHAR,
     regexp_filter VARCHAR,
-    min_size INTEGER,
-    max_size INTEGER
+    min_size FLOAT,
+    max_size FLOAT
 );
 
 CREATE TABLE episodes (
